+++ livekit-plugins/install_local.sh
@@ -20,4 +20,5 @@ pip install \
   "${SCRIPT_DIR}/livekit-plugins-playai" \
   "${SCRIPT_DIR}/livekit-plugins-silero" \
   "${SCRIPT_DIR}/livekit-plugins-turn-detector" \
-  "${SCRIPT_DIR}/livekit-plugins-rime"
+  "${SCRIPT_DIR}/livekit-plugins-rime" \
+  "${SCRIPT_DIR}/livekit-plugins-aws"