+++ livekit-plugins/livekit-plugins-aws/livekit/plugins/aws/log.py
@@ -1,0 +1,5 @@
+import logging
+
+logger = logging.getLogger("livekit.plugins.aws")
+for logger_name in ["botocore", "aiobotocore"]:
+    logging.getLogger(logger_name).setLevel(logging.INFO)