+++ tests/test_llm.py
@@ -9,7 +9,7 @@ from typing import Annotated, Callable, Literal, Optional, Union
 import pytest
 from livekit.agents import APIConnectionError, llm
 from livekit.agents.llm import ChatContext, FunctionContext, TypeInfo, ai_callable
-from livekit.plugins import anthropic, google, openai
+from livekit.plugins import anthropic, aws, google, openai
 from livekit.rtc import VideoBufferType, VideoFrame
 
 
@@ -101,6 +101,7 @@ LLMS: list[Callable[[], llm.LLM]] = [
     pytest.param(lambda: anthropic.LLM(), id="anthropic"),
     pytest.param(lambda: google.LLM(), id="google"),
     pytest.param(lambda: google.LLM(vertexai=True), id="google-vertexai"),
+    pytest.param(lambda: aws.LLM(), id="aws"),
 ]
 
 
@@ -128,6 +129,35 @@ async def test_chat(llm_factory: Callable[[], llm.LLM]):
             text += content
 
     assert len(text) > 0
+
+
+@pytest.mark.parametrize("llm_factory", LLMS)
+async def test_llm_chat_with_consecutive_messages(
+    llm_factory: callable,
+):
+    input_llm = llm_factory()
+
+    chat_ctx = ChatContext()
+    chat_ctx.append(
+        text="Hello, How can I help you today?",
+        role="assistant",
+    )
+    chat_ctx.append(text="I see that you have a busy day ahead.", role="assistant")
+    chat_ctx.append(
+        text="Actually, I need some help with my recent order.", role="user"
+    )
+    chat_ctx.append(text="I want to cancel my order.", role="user")
+
+    stream = input_llm.chat(chat_ctx=chat_ctx)
+    collected_text = ""
+    async for chunk in stream:
+        if not chunk.choices:
+            continue
+        content = chunk.choices[0].delta.content
+        if content:
+            collected_text += content
+
+    assert len(collected_text) > 0, "Expected a non-empty response from the LLM chat"
 
 
 @pytest.mark.parametrize("llm_factory", LLMS)
@@ -347,9 +377,7 @@ async def test_tool_choice_options(
     print(calls)
 
     call_names = {call.call_info.function_info.name for call in calls}
-    if tool_choice == "none" and isinstance(input_llm, anthropic.LLM):
-        assert True
-    else:
+    if tool_choice == "none":
         assert call_names == expected_calls, (
             f"Test '{description}' failed: Expected calls {expected_calls}, but got {call_names}"
         )