+++ livekit-plugins/livekit-plugins-deepgram/CHANGELOG.md
@@ -1,5 +1,15 @@
 # livekit-plugins-deepgram
 
+## 0.6.20
+
+### Patch Changes
+
+- fix(deepgram): fix STT keyterm parameter - [#1535](https://github.com/livekit/agents/pull/1535) ([@wdhwg001](https://github.com/wdhwg001))
+
+- use connection pool for deepgram tts - [#1523](https://github.com/livekit/agents/pull/1523) ([@jayeshp19](https://github.com/jayeshp19))
+
+- remove update options from tts synthesis stream - [#1546](https://github.com/livekit/agents/pull/1546) ([@jayeshp19](https://github.com/jayeshp19))
+
 ## 0.6.19
 
 ### Patch Changes